Moved interfaces.py from ./core/util to ./core/ and moved code for interface method decoration from modules.py to interfaces.py. Updated new import path in all modules.
Included abc.abstractmethod in InterfaceMethod class.
You can make this object an abstractmethod by passing an optional flag
to InterfaceMethod.__init__.

The abstract and non-abstract version of this object can be invoked by
using either decorator "abstract_interface_method" or "interface_method"
respectively.
