+++ core/module.py
@@ -26,110 +26,6 @@ from fysom import Fysom  # provides a final state machine
 from collections import OrderedDict
 from enum import Enum
 from qtpy import QtCore
-from functools import partial
-
-
-class InterfaceMethod:
-    """
-    This object can serve as a replacement (via decorator "@interface_method") for qudi interface
-    methods.
-    It is possible to register several callables, each associated to a unique interface class name.
-    The registered callables can all be called using the decorated method name and adding the
-    keyword argument 'interface' to specify which registered interface to use.
-    If the 'interface' keyword is omitted or the interface is not registered, the call will default
-    to using the callable decorated by "@interface_method".
-
-    This object is compatible with the abc.abstractmethod decorator.
-    If it is declared as abstractmethod, you have to register at least one callable with an
-    interface to count as "implemented".
-    """
-
-    _latest_unregistered_instances = dict()
-
-    def __new__(cls, default_callable, obj=None):
-        func_name = default_callable.__name__
-        print(func_name)
-        if func_name not in cls._latest_unregistered_instances:
-            cls._latest_unregistered_instances[func_name] = super(InterfaceMethod, cls).__new__(cls)
-        return cls._latest_unregistered_instances[func_name]
-
-    def __init__(self, default_callable, obj=None):
-        self._obj = obj
-        self._default_callable = default_callable
-        self.registered = dict()
-
-    def __get__(self, obj=None, cls=None):
-        # It is executed when this instance is referenced as a method
-        if obj is not None and obj is not self._obj:
-            self._default_callable = self._default_callable.__get__(obj, cls)
-            for interface in self.registered:
-                self.registered[interface] = self.registered[interface].__get__(obj, cls)
-        self._obj = obj
-        return self
-
-    def __call__(self, *args, **kwargs):
-        return self._default_callable(*args, **kwargs)
-
-    def __getitem__(self, key):
-        if key not in self.registered:
-            raise KeyError('No method registered for interface "{0}".')
-        return self.registered[key]
-
-    def register(self, interface):
-        """
-        Decorator to register a callable to be used as overloaded function associated with a given
-        interface class name <interface>.
-
-        Example usage in a hardware module:
-
-            @MyInterfaceClass.my_overloaded_method.register('MyInterfaceClass')
-            def some_arbitrary_name1(self, *args, **kwargs):
-                # Do something
-                return
-
-            @MyInterfaceClass.my_overloaded_method.register('MyOtherInterfaceClass')
-            def some_arbitrary_name2(self, *args, **kwargs):
-                # Do something else
-                return
-
-        @param str interface: Name of the interface class the decorated method is associated to
-        @return: Decorator
-        """
-        def decorator(func):
-            self.registered[interface] = func
-            self.__isabstractmethod__ = False
-            InterfaceMethod._latest_unregistered_instances.pop(func.__name__, None)
-            return func
-        return decorator
-
-    # def __repr__(self):
-    #     # Mock representer of bound default callable (will be bound method).
-    #     return self._default_callable.__repr__()
-
-    # Makes this object compatible with abc.abstractmethod
-    @property
-    def __isabstractmethod__(self):
-        if hasattr(self._default_callable, '__isabstractmethod__'):
-            return self._default_callable.__isabstractmethod__
-        return False
-
-    @__isabstractmethod__.setter
-    def __isabstractmethod__(self, flag):
-        if hasattr(self, '__isabstractmethod__'):
-            self._default_callable.__isabstractmethod__ = bool(flag)
-
-
-def interface_method(func):
-    """
-    Decorator to replace a simple interface method with an InterfaceMethod object instance that can
-    register multiple method implementations by the same name and associate each with an interface.
-    This enables a quasi-overloading of interface methods.
-    Is compatible with abc.abstractmethod.
-
-    @param callable func: The callable to be decorated
-    @return InterfaceMethod: Instance of InterfaceMethod to replace the decorated callable
-    """
-    return InterfaceMethod(func)
 
 
 class StatusVar:
