+++ tools/config_gui/listmods.py
@@ -25,7 +25,7 @@ import fnmatch
 import inspect
 import importlib
 
-from qtpy import QtCore, QtWidgets
+from qtpy import QtCore
 
 
 class Module(QtCore.QObject):
@@ -113,7 +113,7 @@ def find_pyfiles(path):
 
 def check_qudi_modules(filelist):
     from core.module import Base
-    from core.util.interfaces import InterfaceMetaclass
+    from core.interfaces import InterfaceMetaclass
     from gui.guibase import GUIBase
     from logic.generic_logic import GenericLogic
     from logic.generic_task import InterruptableTask, PrePostTask
